requirements.txt
boto3==1.34.31
celery==5.3.6
Django==5.0.7
django-cors-headers==4.3.1
django-crum==0.7.9
django_compressor==4.5
django-environ==0.11.2
django-filter==23.5
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
faker==24.3.0
gunicorn==21.2.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
setuptools==69.0.3
redis==5.0.1
python-decouple==3.8
django-tenants==3.6.1
requests==2.32.3
